requirements.txt
pydantic
pymupdf
pytesseract
pillow
scikit-learn
numpy
streamlit
pyyaml
google-genai
python-dotenv
